Change videos to use video src attribute instead of separate source tag
This makes the "error" event fire when a video-related error occurs. It seems that some types of unsupported videos/encodings cause the browser to silently ignore a source, which is not what we want, but we only need one source per video anyway. (It doesn't appear possible to specify the mime type on the video tag, but that may only be used by browsers to skip unsupported types.)